+++ apps/backend/core/src/main/java/com/schemafy/core/project/application/port/in/CreateProjectCommand.java
@@ -2,7 +2,7 @@ package com.schemafy.core.project.application.port.in;
 
 public record CreateProjectCommand(
     String workspaceId,
-    Long dbVendorId,
+    Integer dbVendorId,
     String name,
     String description,
     String requesterId) {
